# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: COGNISAFE-Q Backend API — AWS Serverless (Cognito + API Gateway + Lambda + DynamoDB + SNS)

servers:
  - url: /
    description: API Gateway base URL (set EXPO_PUBLIC_API_GATEWAY_URL)

tags:
  - name: health
    description: Health operations
  - name: profile
    description: User profile management
  - name: contacts
    description: Emergency contacts CRUD
  - name: trips
    description: Trip persistence
  - name: incidents
    description: Emergency incident escalation

security:
  - CognitoJWT: []

paths:
  /health:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns service health status. No auth required.
      security: []
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      description: Auto-creates profile on first call using JWT email.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List emergency contacts
      description: Returns contacts sorted by priority ascending.
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contact"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create emergency contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /contacts/{contactId}:
    parameters:
      - name: contactId
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update emergency contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Contact"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete emergency contact
      responses:
        "204":
          description: Deleted successfully
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips
      description: Returns last 30 trips, newest first.
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Trip"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create trip
      description: Idempotent — returns existing trip if tripId already exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Trip"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /trips/{tripId}:
    parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get specific trip
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Trip"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip (end trip)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripRequest"
      responses:
        "200":
          description: Trip updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Trip"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      description: Returns last 50 incidents, newest first.
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incident"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Escalate emergency incident
      description: |
        Receives confirmed emergency from mobile device.
        - Persists incident to DynamoDB
        - Queries enabled emergency contacts
        - DEMO_MODE=true: logs to CloudWatch
        - DEMO_MODE=false: sends SNS SMS to each contact
        - Idempotent via clientIncidentId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Incident created and notifications sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Incident"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /incidents/{incidentId}:
    parameters:
      - name: incidentId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get specific incident
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Incident"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito ID Token from signIn()

  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Resource belongs to another user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        service:
          type: string
        version:
          type: string
        timestamp:
          type: string
        uptimeMs:
          type: number
        region:
          type: string
      required: [status, service, version, timestamp]

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]

    ProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            userId:
              type: string
            email:
              type: string
            name:
              type: string
              nullable: true
            phone:
              type: string
              nullable: true
            createdAt:
              type: string

    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20

    Contact:
      type: object
      properties:
        contactId:
          type: string
        userId:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        relationship:
          type: string
        priority:
          type: integer
        enabled:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [contactId, userId, name, phone, priority, enabled]

    CreateContactRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        phone:
          type: string
          minLength: 7
          maxLength: 20
        email:
          type: string
          format: email
        relationship:
          type: string
          maxLength: 50
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 1
        enabled:
          type: boolean
          default: true
      required: [name, phone]

    Trip:
      type: object
      properties:
        tripId:
          type: string
        userId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: number
          description: Seconds
        distance:
          type: number
          description: Kilometres
        alertCount:
          type: integer
        emergencyTriggered:
          type: boolean
        safeWindows:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [tripId, userId, startTime]

    CreateTripRequest:
      type: object
      properties:
        tripId:
          type: string
          description: Mobile sessionId — used as DynamoDB PK (idempotency key)
        startTime:
          type: string
          format: date-time
      required: [tripId, startTime]

    UpdateTripRequest:
      type: object
      properties:
        endTime:
          type: string
          format: date-time
        duration:
          type: number
        distance:
          type: number
        alertCount:
          type: integer
        emergencyTriggered:
          type: boolean
        safeWindows:
          type: integer

    Incident:
      type: object
      properties:
        incidentId:
          type: string
        clientIncidentId:
          type: string
          description: Mobile-provided idempotency key
        userId:
          type: string
        tripId:
          type: string
        status:
          type: string
          enum: [EMERGENCY, ALERT, RESOLVED]
        confidence:
          type: number
          minimum: 0
          maximum: 1
        latitude:
          type: number
        longitude:
          type: number
        accuracy:
          type: number
        speed:
          type: number
        heading:
          type: number
        detectedAt:
          type: string
          format: date-time
          description: When mobile detected the incident
        createdAt:
          type: string
          format: date-time
          description: When backend received the incident
        notificationStatus:
          type: string
          enum: [PROCESSING, SENT, FAILED, NO_CONTACTS, DEMO]
        notificationTimestamp:
          type: string
        notifiedContacts:
          type: integer
      required: [incidentId, userId, status, confidence, detectedAt, createdAt, notificationStatus]

    CreateIncidentRequest:
      type: object
      properties:
        tripId:
          type: string
        clientIncidentId:
          type: string
          description: Idempotency key — reuse mobile sessionId+timestamp
        status:
          type: string
          enum: [EMERGENCY, ALERT]
          default: EMERGENCY
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: On-device ML confidence score (0.0–1.0)
        location:
          type: object
          properties:
            latitude:
              type: number
              minimum: -90
              maximum: 90
            longitude:
              type: number
              minimum: -180
              maximum: 180
            accuracy:
              type: number
            speed:
              type: number
            heading:
              type: number
          required: [latitude, longitude]
        detectedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when mobile detected the incident
      required: [confidence, detectedAt]